requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
requests==2.31.0
redis==5.0.1
eventlet==0.33.3
gevent==23.9.1
gevent-websocket==0.10.1
numpy==1.26.2
pytest==7.4.3
pytest-cov==4.1.0
gunicorn==21.2.0 